::foreach: Prevent an "N+1 query problem" with thumbnails
post-thumbnail-template.php has code that goes like

 		if ( in_the_loop() )
			update_post_thumbnail_cache();

As a result, thumbnails are only properly 1+1-queried from within the
main loop. (This is unlike posts and comments.) Iterating over all
things that have a thumbnail (e.g. in a faculty gallery) therefore
costs N queries to the database, unless we find a way to override the
return value of in_the_loop() in the snippet above.

* In ::foreach, set global $wp_query to be the query that returns all
  objects to iterate over

* Set that query's ->in_the_loop to true

* Undo all of the above in a finally block

* Encapsulate all of the above in a nice, coyly-documented
  _InTheLoopHelper class.